Add OpenStack OLM subscription configurations
Ability to install a specific version of the operator(s) when multiple
versions are available in the OLM catalog.

Added new kustomization files and YAML configurations for OpenStack
operator subscriptions using OLM. This includes catalog sources,
namespaces, and operator groups for openstack-operators, as well as
subscription definitions for the OpenStack components when the version
is below v1.0.7.

Kustomization facilitates version-specific configurations starting
with v1.0.3 and v1.0.6, incorporating modifications like the
elimination of the AnsibleEE operator in subsequent versions. From
version 1.0.7 onwards, only the OpenStack operator is retained.

All configurations reference a central ConfigMap for dynamic value
replacements.